+++ lib/components/organisms/BadgeTile/styles.tsx
@@ -1,4 +1,5 @@
 import { StyleSheet } from 'react-native';
+import { IS_WEB } from '../../../constants/device';
 import { useWllSdk } from '../../../context/WllSdkContext';
 import { useResponsive } from '../../../hooks/useResponsive';
 import { useResponsiveValue } from '../../../utils/responsiveHelper';
@@ -44,6 +45,8 @@ export const useBadgeTileStyles = (): ReturnType<typeof StyleSheet.create> => {
       fontWeight: 'bold',
     },
     header: {
+      width: '100%',
+      ...(!IS_WEB && { flexBasis: '50%' }),
       alignItems: 'center',
       justifyContent: 'center',
       position: 'relative',