+++ lib/components/organisms/VenueTile/venue-tile-media.tsx
@@ -1,14 +1,10 @@
 import React from 'react';
 import { StyleProp, View, ViewStyle } from 'react-native';
-import { VenueTileConfig } from '../../../types/tile';
+import { TileHeight, VenueTileConfig } from '../../../types/tile';
 import { isContextValid } from '../../../utils/contextHelpers';
 import { Icon, ProgressiveImage } from '../../atoms';
 import { useTileContext } from '../../atoms/BaseTile';
 import { useVenueTileStyles } from './styles';
-
-type VenueTileMediaProps = {
-  isArtworkOnly?: boolean;
-};
 
 // We are using percentage values for flexBasis, which is valid in
 // React Native but TypeScript expects a number.
@@ -20,14 +16,10 @@ type ResponsiveViewStyle = StyleProp<ViewStyle> & {
 /**
  * Renders the media for a Venue Tile.
  *
- * @param props {VenueTileMediaProps} - Component props
- * @param props.isArtworkOnly {boolean} - Whether the media should be rendered as an artwork only component
  * @returns React.ReactElement or null if no artwork URL is present or artwork should not be shown
  */
 
-export const VenueTileMedia = ({
-  isArtworkOnly = false,
-}: VenueTileMediaProps): React.ReactElement | null => {
+export const VenueTileMedia = (): React.ReactElement | null => {
   const styles = useVenueTileStyles();
   const tileContext = useTileContext();
 
@@ -39,11 +31,12 @@ export const VenueTileMedia = ({
     isLocked = false,
   } = tileContext.configuration as VenueTileConfig;
 
-  if (!artworkUrl) return null;
-
   const containerStyle: ResponsiveViewStyle = {
-    flexBasis: isArtworkOnly ? '100%' : '50%',
+    flexBasis: '50%',
+    ...(tileContext.tileHeight === TileHeight.Half && { minHeight: 0 }),
   };
+
+  const showCenteredPin = !artworkUrl && !isLocked;
 
   return (
     <View
@@ -52,16 +45,23 @@ export const VenueTileMedia = ({
       role="img"
       accessibilityLabel={`Venue image for ${name}`}
     >
-      <ProgressiveImage
-        source={{ uri: artworkUrl }}
-        style={styles.image}
-        alt={`Venue image for ${name}`}
-      />
-      {isLocked && (
-        <View style={styles.lockOverlay} testID="lock-overlay">
-          <Icon name="LockKeyhole" size={48} color="white" />
-        </View>
+      {artworkUrl && (
+        <ProgressiveImage
+          source={{ uri: artworkUrl }}
+          style={styles.image}
+          alt={`Venue image for ${name}`}
+        />
       )}
+      <View style={styles.lockOverlay} testID="lock-overlay">
+        {isLocked && <Icon name="LockKeyhole" size={48} color="white" />}
+        {showCenteredPin ? (
+          <Icon name="MapPin" size={48} color="white" />
+        ) : (
+          <View style={styles.pinOverlay} testID="pin-overlay">
+            <Icon name="MapPin" size={24} color="white" />
+          </View>
+        )}
+      </View>
     </View>
   );
 };